Lesson 3 homeworks 4rd part

+++ src/main/java/com/javaguru/student_deniss_lohins/lesson_3/level_6/BankAccount.java
@@ -1,5 +1,7 @@
 package com.javaguru.student_deniss_lohins.lesson_3.level_6;
 
+import java.math.BigDecimal;
+
 /*
         Найти логическую ошибку в коде.
         Исправить и запустить программу так, что бы она выводила
@@ -10,7 +12,7 @@ class BankAccount {
 
     private String ownerFirstName;
     private String owner;
-    private int money;
+    private BigDecimal money;
 
     public BankAccount(String owner, int moneyAmount) {
         this.owner = owner;
@@ -21,7 +23,7 @@ class BankAccount {
         return this.owner;
     }
 
-    public int getMoney() {
+    public BigDecimal getMoney() {
         return this.money;
     }
 
@@ -32,7 +34,7 @@ class BankAccountDemo {
     public static void main(String[] args) {
         BankAccount bankAccount = new BankAccount("Viktor", 1000000);
         String ownerFirstName = bankAccount.getOwner();
-        int moneyAmount = bankAccount.getMoney();
+        BigDecimal moneyAmount = bankAccount.getMoney();
         System.out.println("Owner = " + ownerFirstName);
         System.out.println("Money = " + moneyAmount);
     }
